# 09_小程序需求/05_API文档/openapi_v1.0.yaml
openapi: 3.0.3
info:
  title: Bark & Bar 微信小程序 API
  version: 1.0.0
  description: |
    Bark & Bar 系统 V1.0 开发基线接口。该文件覆盖首期核心业务动作，
    具体字段以 PRD V1.1、数据库设计 V1.1 和业务规则文档为准。
servers:
  - url: https://api.example.com/api/v1
    description: production placeholder
  - url: https://staging-api.example.com/api/v1
    description: staging placeholder
...
security:
  - bearerAuth: []
paths:
...
  /customer/reservation-slots:
    get:
      tags: [CustomerReservations]
      summary: 查询可预约时段
      parameters:
        - name: date
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        '200':
          description: 可预约时段
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationSlotsResponse'
  /customer/reservations:
    get:
      tags: [CustomerReservations]
      summary: 查询我的预约
      parameters:
        - name: user_id
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 我的预约列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reservation'
    post:
      tags: [CustomerReservations]
      summary: 创建预约
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReservationRequest'
      responses:
        '200':
          description: 预约创建结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateReservationResponse'
  /customer/reservations/{reservation_id}/cancel:
    post:
      tags: [CustomerReservations]
      summary: 取消预约或申请改期
      parameters:
        - $ref: '#/components/parameters/ReservationId'
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, reason]
              properties:
                user_id: { type: string }
                reason: { type: string }
                request_reschedule: { type: boolean }
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /customer/reservations/{reservation_id}/checkin-code:
    post:
      tags: [CustomerReservations]
      summary: 生成60秒到店核验码
      parameters:
        - $ref: '#/components/parameters/ReservationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id: { type: string }
      responses:
        '200':
          description: 到店核验码
          content:
            application/json:
              schema:
                type: object
                properties:
                  reservation_id: { type: string }
                  code: { type: string }
                  expires_in_seconds: { type: integer }
                  expires_at: { type: string, format: date-time }
...
  /admin/reservations:
    get:
      tags: [AdminReservations]
      summary: 查询员工端预约列表
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [PENDING_REVIEW, CONFIRMED, CANCELED, RESCHEDULE_REQUESTED, REJECTED, CHECKED_IN]
        - name: staff_pin
          in: query
          required: false
          schema: { type: string }
        - name: staff_token
          in: query
          required: false
          description: 本地员工 token；优先使用。旧 staff_pin 暂保留兼容。
          schema: { type: string }
      responses:
        '200':
          description: 预约列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reservation'
  /admin/reservations/{reservation_id}/review:
    post:
      tags: [AdminReservations]
      summary: 员工审核预约或改期申请
      parameters:
        - $ref: '#/components/parameters/ReservationId'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [APPROVE, REJECT]
                staff_pin: { type: string }
                staff_token: { type: string }
                note: { type: string }
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /admin/reservations/{reservation_id}/checkin:
    post:
      tags: [AdminReservations]
      summary: 员工确认到店核验
      parameters:
        - $ref: '#/components/parameters/ReservationId'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                staff_pin: { type: string }
                staff_token: { type: string }
                note: { type: string }
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /admin/reservations/verify-code:
    post:
      tags: [AdminAdmissions]
      summary: 员工用动态核验码确认到店
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                staff_pin: { type: string }
                staff_token: { type: string }
                code: { type: string }
                note: { type: string }
      responses:
        '200':
          $ref: '#/components/responses/Ok'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RequestId:
      name: X-Request-Id
      in: header
      required: true
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string }
...
    ReservationId:
      name: reservation_id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Ok:
      description: 操作成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
...
  schemas:
    Envelope:
      type: object
      required: [request_id, data]
      properties:
        request_id: { type: string }
        data:
          type: object
...
    ReservationSlotsResponse:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: array
              items:
                type: object
                properties:
                  slot_id: { type: string }
                  starts_at: { type: string, format: date-time }
                  ends_at: { type: string, format: date-time }
                  remaining_people: { type: integer }
                  remaining_dogs: { type: integer }
                  requires_manual_review: { type: boolean }
    Reservation:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        slotId: { type: string }
        peopleCount: { type: integer }
        petIds:
          type: array
          items: { type: string }
        entryType: { type: string }
        status:
          type: string
          enum: [PENDING_REVIEW, CONFIRMED, CANCELED, RESCHEDULE_REQUESTED, REJECTED, CHECKED_IN]
        estimatedAmountFen: { type: integer }
        reviewReason: { type: string }
        adminReviewNote: { type: string }
        reviewedAt: { type: string, format: date-time }
        checkinCodeExpiresAt: { type: string, format: date-time }
        cancelReason: { type: string }
        rescheduleRequested: { type: boolean }
        canceledAt: { type: string, format: date-time }
        checkedInAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
    CreateReservationRequest:
      type: object
      required: [slot_id, people_count, pet_ids, entry_type]
      properties:
        slot_id: { type: string }
        people_count: { type: integer, minimum: 1 }
        pet_ids:
          type: array
          items: { type: string }
        entry_type:
          type: string
          enum: [single_ticket, ten_pass_card, season_card, annual_card]
        card_id: { type: string }
        extra_dog_payment_choice:
          type: string
          enum: [deduct_card_unit, pay_extra_fee]
    CreateReservationResponse:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                reservation_id: { type: string }
                status: { type: string }
                payment_required: { type: boolean }
                payment_order_id: { type: string }
                review_required: { type: boolean }
                benefit_locked: { type: boolean }
...